Refactor Listings and Offers tables to use ControlledTable component. Replace OrdersTable with ControlledTable in ListingsTable and OffersTable.
