+++ playgrounds/react-vite/src/components/collectible/tables/ListingsTable.tsx
@@ -1,5 +1,5 @@
-import { Box, Text, useToast } from '@0xsequence/design-system';
-import { type Order, compareAddress } from '@0xsequence/marketplace-sdk';
+import { Box, GradientAvatar, Text, useToast } from '@0xsequence/design-system';
+import { type Order, compareAddress, truncateMiddle } from '@0xsequence/marketplace-sdk';
 import {
 	useBuyModal,
 	useCancelOrder,
@@ -9,7 +9,10 @@ import {
 import { useState } from 'react';
 import { useAccount } from 'wagmi';
 import { useMarketplace } from '../../../lib/MarketplaceContext';
-import { OrdersTable } from './OrdersTable';
+import { ControlledTable, type Column } from '../../../lib/Table/ControlledTable';
+import { CurrencyCell } from './CurrencyCell';
+import toTitleCaseFromSnakeCase from '../../../lib/util/toTitleCaseFromSnakeCase';
+import { ActionCell } from './ActionCell';
 
 export const ListingsTable = () => {
 	const { collectionAddress, chainId, collectibleId } = useMarketplace();
@@ -103,6 +106,70 @@ export const ListingsTable = () => {
 		}
 	};
 
+	const columns: Column<Order>[] = [
+		{
+			header: 'Price',
+			key: 'priceAmountFormatted',
+			render: (order) => (
+				<Text fontFamily="body" color="text100">
+					{order.priceAmountFormatted}
+				</Text>
+			),
+		},
+		{
+			header: 'Currency',
+			key: 'priceCurrencyAddress',
+			render: (order) => (
+				<CurrencyCell currencyAddress={order.priceCurrencyAddress} />
+			),
+		},
+		{
+			header: 'Seller',
+			key: 'createdBy',
+			render: (order) => (
+				<Box display="flex" alignItems="center" gap="1">
+					<GradientAvatar address={order.createdBy} size="xs" />
+					<Text fontFamily="body" color="text100">
+						{truncateMiddle(order.createdBy, 3, 4)}
+					</Text>
+				</Box>
+			),
+		},
+		{
+			header: 'Expiration',
+			key: 'validUntil',
+			render: (order) => (
+				<Text fontFamily="body" color="text100">
+					{new Date(order.validUntil).toLocaleDateString()}
+				</Text>
+			),
+		},
+		{
+			header: 'Orderbook',
+			key: 'marketplace',
+			render: (order) => (
+				<Box
+					background="backgroundMuted"
+					paddingX="2"
+					paddingY="1"
+					display="inline-block"
+					borderRadius="xs"
+				>
+					<Text fontSize="xsmall" fontFamily="body" fontWeight="bold">
+						{toTitleCaseFromSnakeCase(order.marketplace)}
+					</Text>
+				</Box>
+			),
+		},
+		{
+			header: 'Actions',
+			key: 'actions',
+			render: (order) => (
+				<ActionCell order={order} getLabel={getLabel} onAction={handleAction} />
+			),
+		},
+	];
+
 	return (
 		<>
 			<Box
@@ -121,18 +188,18 @@ export const ListingsTable = () => {
 				</Text>
 			</Box>
 
-			<OrdersTable
+			<ControlledTable<Order>
 				isLoading={listingsLoading}
 				items={listings?.listings}
+				columns={columns}
 				emptyMessage="No listings available"
-				getLabel={getLabel}
-				onAction={handleAction}
-				nextPage={() => setPage((prev) => prev + 1)}
-				prevPage={() => setPage((prev) => prev - 1)}
-				isPrevDisabled={page <= 1}
-				isNextDisabled={!listings?.page?.more}
-				type="listings"
-				currentPage={page}
+				pagination={{
+					onNextPage: () => setPage((prev) => prev + 1),
+					onPrevPage: () => setPage((prev) => prev - 1),
+					isPrevDisabled: page <= 1,
+					isNextDisabled: !listings?.page?.more,
+					currentPage: page,
+				}}
 			/>
 		</>
 	);
